Added support for detection and loading of simulation backends
By default the Feature's autodetection tries to load a production
backend first and if not available falls back to a simulation backend.

This behavior can be controlled by the AutoDiscoveryMode. Which
backend type got loaded can be checked with the AutoDiscoveryResult.

The QtIVIServiceManager got improved by adding a SearchFlag to the
find method, which controls what type of backends should be included

Change-Id: Ife3bb3672f21d2fb5b2d7236099de1df17487530
Reviewed-by: Johan Thelin <johan.thelin@pelagicore.com>