Doc: Fixes for linking and table of contents
The documentation build process for Qt Automotive Suite currently
does not have access to other doc sets that we are trying to link
to. As a workaround, use external URLs for linking.

Also, rename the landing page to index.html and list only the
Overview page as a TOC subsection, to avoid duplicating a link to
the index page.

Change-Id: I7501bb12b09f31185ff134fd7629f9aaf3d70b5e
Reviewed-by: Risto Avila <risto.avila@qt.io>