ivigenerator: Fix the Makefile errors when running inside QtCreator
The qmake_all step is already trying to call the generator
even though it is not build yet. To fix the problem a dependency
to a dummy file is created, which is touched once the generator
is ready to be used. This will cause a rerun of the qmake for
all project files which use the ivigenerator feature.

Change-Id: I5a421597457dccc29ab48811a9d0588c04c5cf2b
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>