Handle building modules properly
Extra steps are needed for syncqt to find the headers. After the
generation of the frontend files, the headers are copied in the
src folder in qtbase.

When building a module, rather than using CONFIG += ivigenerator,
simply define the required qface and module related variables and
load(ivigenerator_qt_module)

This also handles the QFACE_MODULE_NAME variable: the generator has a
--module option which optionaly indicates the name of the Qt Module
that is generated by the project.

This is needed by the backend templates to know which include paths to
use to find the frontend base interface class.

This also changes the generated files so that the generated .pri
has the same base name (in lower case) as the input qface file.
And updates the generator version to 2.0

Change-Id: I9d2716255affcb5c6f319a6f23202ae3bb6c115c
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>