Fix for overriding zoned features
Initialize the pointer to the service object for zoned child
features (doesn’t do the full registration as it doesn’t need
to connect to change requests). This is needed to be
able to identify all instances of front end classes which
are talking to the same backend.

Also always update the client side value even if overriding is on.

Change-Id: I2d756106c14536936451f61a502a13a1bb28e8dd
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>