Improve Qt AS module and tool descriptions on Overview page
Clarified the description of some Qt Automotive Suite tools and
and related modules. Added also a high-level picture of Qt AS
structure. The picture is based on the blog post written by
Alistair Adams in June 8th, 2016. The picture is updated
according to Qt brand guidelines (font, colors).

Task-number: QTAUTO-558
Change-Id: If5828fe307a1f99cb8c29d36a6e18805015fbe42
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>