ivigenerator: Improve the deploy-virtualenv.sh script
On some linux systems the python executable is linked to a libpython.so
file, which is not put into the virtualenv. This file needs to be copied
as well and made sure that it is found when the binary is executed.

Also the system module directory is not always located in <prefix>/lib.
To find the correct locations we use the python sys module to print
out all paths and search for the files in all starting with <prefix>

Task-number: QTAUTO-289
Change-Id: I9b1077c89501803fa83dab02e8238fa48a005298
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>