Add the QIviPendingReply class
The QIviPendingReply class is templated to provide a typesafe C++ API, but
at the same time being usable from QML directly.
For C++, signals are provided to inform the user that the reply has succeeded
or failed. From QML one can either use the signals or using the "then"
method to register JavaScript callbacks very similar to a JavaScript Promise.
As the class provides the "then" method, it should also be convertable into
a real JavaScript Promise.

Task-number: QTAUTO-836
Change-Id: I030c70a964b48a5083ecef7f527d8a0b1c85f137
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>