Fix the custom signals and slots in qtro templates
Fix the custom slot implememtation in qtro templates. Calling a slot
on a remote object replica causes an asynchronous call and returns a
pending result object. The current backend slot implementation merely
waits for the call to finish before returning to the code that uses it.
This behavior will change when all method calls are made asynchronous.
Also add tests for passing a remote objects initial values to newly
connected clients.

Change-Id: I278bfc583a657e8a30265af8ec6c02e178f0f245
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>