Add the QIviPagingModel class
The QIviPagingModel is reusing the code from the QIviSearchAndBrowseModel
and implements only the interface for retrieving the content of a model
using "Paging".

This commit just adds the new class. New functionality will be added
in following commits, as well as using the new class as the base
for QIviSearchAndBrowseModel.

Task-number: AUTOSUITE-420
Change-Id: I45f068efcb1d76e566556ee6ff54f429dc6b28da
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>