Add support for additional annotation files
The ivigenerator as well as its qmake integration now support to pass
additional annotation files. This can be useful when annotations only
need to be applied for a specific part, e.g. the backend plugin.

This change also makes use of the feature in the autotests which now
use a no-private.yaml annotation file for testing the same qface file
as we use for our normal generation tests.

Task-number: QTAUTO-848
Change-Id: Ib626775d2339ebf2002bc689da8303e7caf6edfa
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>