Add new (un)registerInstance functions to the QIviPagingModelInterface
These functions help to maintain the state in the backend
implementations. The registerInstance function is called for
every new instance of a model using this backend. The
unregisterInstance is called once the model instance is
destroyed and the all data for this model can be cleanup
in the backend as well.

Task-number: AUTOSUITE-435
Change-Id: I2390c4fd423acb40cd50c6a2d2ff22c15aef15ab
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>