Add a generic backend() function to QIviAbstractFeaturePrivate
Using this function in the derived classes makes the code much cleaner
and less error-prone as always interfaceName() is used to get the
backend object, which makes sure deriving the interface works as well.

Also get rid of the disconnectServiceObject reimplementation where
possible and document when it makes sense to reimplement this function.

Change-Id: I05ade6e8ca8d3829cabcbbdf312db4b46a25044d
Reviewed-by: Antti Hölttä <ahoelttae@luxoft.com>