Add support for overriding the simulation/data-file at runtime
The QIviSimulationEngine can now take an extra identifier in the constructor
which is used to identify override settings for this simulation engine.

Using the following environment variable schema it is possible to set
and override which the simulation engine will use instead of the compiled
in simulation/data-file.

QTIVI_SIMULATION_OVERRIDE=<identifier>=<QML-file>
QTIVI_SIMULATION_DATA_OVERRIDE=<identifier>=<JSON-file>

Change-Id: Id2a921d8b74c619ea9662d413f0a7296de949b91
Fixes: AUTOSUITE-703
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>