Show a meaningful error message debug and release libraries mixed
On mac it can easily happen that debug and release libraries are
mixed when building an application. E.g. using one of the qface
examples. In general qtivicore tries to load the correct plugin version
if available, but if the wrong one was loaded or the appropriate one
was not available it can happen that plugin interfaces can't be
casted to the correct types and thus the class doesn't work.
When this happens a more descriptive error message is now shown and
the backend is not used.

Change-Id: I4ae63ac87509ec3feb1ae65750d8e0ee71ad6520
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>