ivigenerator: Don't use the javadoc syntax in our qface files
In theory the documentation inside the qface files could be
transformed using the javadoc syntax and the parse_doc filter
function, but so far we only supported the brief tag and already
used qdoc commands in the function directly. As the tag functionality
of parse_doc is nothing we can really use, better use qdoc commands
in the qface documentation right away.

Change-Id: I9b23fe84acc6cd942e6fa7ec5933911b84ce22ec
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>