Make it more convenient to use a QIviPendingReply from C++
The templated version now provides a value constructor which sets the
reply to success and stores the value. This is the counter part to the
static createFailedPendingReply method.

In addition a 'then' method has been added which can take two C++
lambda functions and is more convenient than using connect on the
watcher object.

Change-Id: I009e5034da178e85ed93a5bb7759a76afb6eb864
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>