backend_qtro: Create the connection in initialize() not before
Move the complete connection handling to a new connectToNode()
function, which always reads the settings file and checks for
a change of the connection URL and reconnect if needed.

This function is called in initialize() to make it possible
to change the connection URL in the settings file and reassign
the backend to reconnect.

This should also improve the performance as the actual connection
is created when it is needed for the first time instead of already
connecting in the backen constructor

Change-Id: Iae7f21c19dad87f198181b8f39d5a6d08437f76c
Fixes: AUTOSUITE-1000
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>