Move the complete QML type registration to the QtIviCore Module class
We do it like this already in the autogenerated module classes and it
maintaining such a global registerQmlTypes function makes the type
registration in the plugin pretty easy and also makes it possible for
an app developer to do the same registration without a plugin.

In addition move the QIviServiceObject* type registration into the
QIviServiceManager to make sure the type is registered once it is
instantiated.

Change-Id: I12a7e797f6c6a35cf12988eccd2e6fc2ccb7367d
Reviewed-by: Antti Hölttä <ahoelttae@luxoft.com>